Initial Commit for USB CDC example projet
This project serves as an example CDC and a test run for using github with
Eclipse CDT.  The project is built for the ST32F4 Discovery board and
enumerates the device as a virtual COM port.  Simple communications can be
used to toggle LEDs.

All project and system files are included with this upload as we do not
know yet the minimal upload required to share projects.